Move queue sync functionality to separate file
